refactor: move profit trader and path finderlogic to separate file
